backend/requirements.txt
Django==3.2.15
djangorestframework==3.12.4
Pillow==9.2.0
djoser==2.1.0
django-filter==22.1